move more parse code to libcomp
