+++ qucs/qucs-lib/libcomp.cpp
@@ -16,6 +16,7 @@
  ***************************************************************************/
 
 #include "libcomp.h"
+#include <cmath>
 #include <QPainter>
 
 void QucsLibComponent::draw(QWidget& w) const
@@ -264,4 +265,213 @@ QucsLibComponent::QucsLibComponent(const QString& Lib_, const QString& Comp_)
   // resize(dx, y2-y1 + TextHeight+4); FIXME: resize what?!
   // update(); fixme: update who?
   PortNumber = PortNo;
+}
+
+int QucsLibComponent::analyseLine(const QString& Row)
+{
+  QPen Pen;
+  QBrush Brush;
+  QColor Color;
+  QString s;
+  int i1, i2, i3, i4, i5, i6;
+
+  s = Row.section(' ',0,0);    // component type
+  if(s == ".PortSym") {  // here: ports are open nodes
+    if(!getCompLineIntegers(Row, &i1, &i2, &i3))  return -1;
+    Arcs.append(new struct Arc(i1-4, i2-4, 8, 8, 0, 16*360,
+                               QPen(Qt::red,1)));
+
+    if((i1-4) < x1)  x1 = i1-4;  // keep track of component boundings
+    if((i1+4) > x2)  x2 = i1+4;
+    if((i2-4) < y1)  y1 = i2-4;
+    if((i2+4) > y2)  y2 = i2+4;
+    return 0;   // do not count Ports
+  }
+  else if(s == "Line") {
+    if(!getCompLineIntegers(Row, &i1, &i2, &i3, &i4))  return -1;
+    if(!getPen(Row, Pen, 5))  return -1;
+    i3 += i1;
+    i4 += i2;
+    Lines.append(new Line(i1, i2, i3, i4, Pen));
+
+    if(i1 < x1)  x1 = i1;  // keep track of component boundings
+    if(i1 > x2)  x2 = i1;
+    if(i2 < y1)  y1 = i2;
+    if(i2 > y2)  y2 = i2;
+    if(i3 < x1)  x1 = i3;
+    if(i3 > x2)  x2 = i3;
+    if(i4 < y1)  y1 = i4;
+    if(i4 > y2)  y2 = i4;
+    return 1;
+  }
+  else if(s == "EArc") {
+    if(!getCompLineIntegers(Row, &i1, &i2, &i3, &i4, &i5, &i6))  return -1;
+    if(!getPen(Row, Pen, 7))  return -1;
+    Arcs.append(new struct Arc(i1, i2, i3, i4, i5, i6, Pen));
+
+    if(i1 < x1)  x1 = i1;  // keep track of component boundings
+    if(i1+i3 > x2)  x2 = i1+i3;
+    if(i2 < y1)  y1 = i2;
+    if(i2+i4 > y2)  y2 = i2+i4;
+    return 1;
+  }
+  else if(s == ".ID") {
+    if(!getCompLineIntegers(Row, &i1, &i2))  return -1;
+    Text_x = i1;
+    Text_y = i2;
+    Prefix = Row.section(' ',3,3);
+    if(Prefix.isEmpty())  Prefix = "X";
+    return 0;   // do not count IDs
+  }
+  else if(s == "Arrow") {
+    if(!getCompLineIntegers(Row, &i1, &i2, &i3, &i4, &i5, &i6))  return -1;
+    if(!getPen(Row, Pen, 7))  return -1;
+
+    double beta   = atan2(double(i6), double(i5));
+    double phi    = atan2(double(i4), double(i3));
+    double Length = sqrt(double(i6*i6 + i5*i5));
+
+    i3 += i1;
+    i4 += i2;
+    if(i1 < x1)  x1 = i1;  // keep track of component boundings
+    if(i1 > x2)  x2 = i1;
+    if(i3 < x1)  x1 = i3;
+    if(i3 > x2)  x2 = i3;
+    if(i2 < y1)  y1 = i2;
+    if(i2 > y2)  y2 = i2;
+    if(i4 < y1)  y1 = i4;
+    if(i4 > y2)  y2 = i4;
+
+    Lines.append(new Line(i1, i2, i3, i4, Pen));   // base line
+
+    double w = beta+phi;
+    i5 = i3-int(Length*cos(w));
+    i6 = i4-int(Length*sin(w));
+    Lines.append(new Line(i3, i4, i5, i6, Pen)); // arrow head
+    if(i5 < x1)  x1 = i5;  // keep track of component boundings
+    if(i5 > x2)  x2 = i5;
+    if(i6 < y1)  y1 = i6;
+    if(i6 > y2)  y2 = i6;
+
+    w = phi-beta;
+    i5 = i3-int(Length*cos(w));
+    i6 = i4-int(Length*sin(w));
+    Lines.append(new Line(i3, i4, i5, i6, Pen));
+    if(i5 < x1)  x1 = i5;  // keep track of component boundings
+    if(i5 > x2)  x2 = i5;
+    if(i6 < y1)  y1 = i6;
+    if(i6 > y2)  y2 = i6;
+
+    return 1;
+  }
+  else if(s == "Ellipse") {
+    if(!getCompLineIntegers(Row, &i1, &i2, &i3, &i4))  return -1;
+    if(!getPen(Row, Pen, 5))  return -1;
+    if(!getBrush(Row, Brush, 8))  return -1;
+    Ellips.append(new Area(i1, i2, i3, i4, Pen, Brush));
+
+    if(i1 < x1)  x1 = i1;  // keep track of component boundings
+    if(i1 > x2)  x2 = i1;
+    if(i2 < y1)  y1 = i2;
+    if(i2 > y2)  y2 = i2;
+    if(i1+i3 < x1)  x1 = i1+i3;
+    if(i1+i3 > x2)  x2 = i1+i3;
+    if(i2+i4 < y1)  y1 = i2+i4;
+    if(i2+i4 > y2)  y2 = i2+i4;
+    return 1;
+  }
+  else if(s == "Rectangle") {
+    if(!getCompLineIntegers(Row, &i1, &i2, &i3, &i4))  return -1;
+    if(!getPen(Row, Pen, 5))  return -1;
+    if(!getBrush(Row, Brush, 8))  return -1;
+    Rects.append(new Area(i1, i2, i3, i4, Pen, Brush));
+
+    if(i1 < x1)  x1 = i1;  // keep track of component boundings
+    if(i1 > x2)  x2 = i1;
+    if(i2 < y1)  y1 = i2;
+    if(i2 > y2)  y2 = i2;
+    if(i1+i3 < x1)  x1 = i1+i3;
+    if(i1+i3 > x2)  x2 = i1+i3;
+    if(i2+i4 < y1)  y1 = i2+i4;
+    if(i2+i4 > y2)  y2 = i2+i4;
+    return 1;
+  }
+  else if(s == "Text") {  // must be last in order to reuse "s" *********
+    if(!getCompLineIntegers(Row, &i1, &i2, &i3))  return -1;
+    Color.setNamedColor(Row.section(' ',4,4));
+    if(!Color.isValid()) return -1;
+
+    s = Row.mid(Row.indexOf('"')+1);    // Text (can contain " !!!)
+    s = s.left(s.length()-1);
+    if(s.isEmpty()) return -1;
+    s.replace("\\n", "\n");
+    s.replace("\\\\", "\\");
+
+    Texts.append(new Text(i1, i2, s, Color, float(i3)));
+
+    QFont Font(QucsSettings.font);
+    Font.setPointSizeF(float(i3));
+    QFontMetrics  metrics(Font, 0);  // use the the screen-compatible metric
+    QSize r = metrics.size(0, s);  // get size of text
+    i3 = i1 + r.width();
+    i4 = i2 + r.height();
+
+    if(i1 < x1)  x1 = i1;  // keep track of component boundings
+    if(i3 > x2)  x2 = i3;
+    if(i2 < y1)  y1 = i2;
+    if(i4 > y2)  y2 = i4;
+  }
+
+  return 0;
+}
+
+
+// ---------------------------------------------------------------------
+bool QucsLibComponent::getPen(const QString& s, QPen& Pen, int i)
+{
+  bool ok;
+  QString n;
+
+  n = s.section(' ',i,i);    // color
+  QColor co;
+  co.setNamedColor(n);
+  Pen.setColor(co);
+  if(!Pen.color().isValid()) return false;
+
+  i++;
+  n = s.section(' ',i,i);    // thickness
+  Pen.setWidth(n.toInt(&ok));
+  if(!ok) return false;
+
+  i++;
+  n = s.section(' ',i,i);    // line style
+  Pen.setStyle((Qt::PenStyle)n.toInt(&ok));
+  if(!ok) return false;
+
+  return true;
+}
+
+// ---------------------------------------------------------------------
+bool QucsLibComponent::getBrush(const QString& s, QBrush& Brush, int i)
+{
+  bool ok;
+  QString n;
+
+  n = s.section(' ',i,i);    // fill color
+  QColor co;
+  co.setNamedColor(n);
+  Brush.setColor(co);
+  if(!Brush.color().isValid()) return false;
+
+  i++;
+  n = s.section(' ',i,i);    // fill style
+  Brush.setStyle((Qt::BrushStyle)n.toInt(&ok));
+  if(!ok) return false;
+
+  i++;
+  n = s.section(' ',i,i);    // filled
+  if(n.toInt(&ok) == 0) Brush.setStyle(Qt::NoBrush);
+  if(!ok) return false;
+
+  return true;
 }
